+++ src/main/java/com/seckill/agent/entity/SeckillGoods.java
@@ -2,15 +2,16 @@ package com.seckill.agent.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 /**
  * @author lr-qixin
  * */
 @Data
 public class SeckillGoods {
-    private Long id;
-    private Long goodsId;
-    private double secondKillPrice;
+    private Integer id;
+    private Integer goodsId;
+    private BigDecimal secondKillPrice;
     private Integer stockCount;
     private Date startDate;
     private Date endDate;
